+++ components/block/LatestBlocks.tsx
@@ -150,7 +150,7 @@ const LatestBlocks = (props: any) => {
     event: React.ChangeEvent<HTMLInputElement>
   ) => {
     setPageSize(+event.target.value);
-    setPage(0);
+    setPage(1);
   };
 
   moment.relativeTimeThreshold("s", 59);
@@ -355,6 +355,13 @@ const LatestBlocks = (props: any) => {
               page={page}
               onChangePage={handleChangePage}
               onChangeRowsPerPage={handleChangeRowsPerPage}
+              backIconButtonProps={{
+                'aria-label': 'Previous',
+                'disabled': page === 1,
+              }}
+              nextIconButtonProps={{
+                'aria-label': 'Next',
+              }}
             />
           ) : null}
         </Paper>