+++ CHANGELOG.md
@@ -34,4 +34,4 @@
 * [#30] Stored the GraphQL & account endpoint in environment variable
 * [#41] Updated Middle Ellipsis screen size function to useMediaQuery 
 * Updated Grid size on medium screens to 7
-
+* Disabled back button on Page 1 in TablePagination